Move Hamiltonian tests
